requirements.txt
Django>=4.2
django-allauth>=0.63
Pillow>=10.0
requests
PyJWT
cryptography
psycopg2-binary

